+++ gradle.properties
@@ -1,8 +1,8 @@
 org.gradle.jvmargs  = -Xmx1G
 
 #Fabric properties
-minecraft_version   = 1.16.2-pre3
-yarn_mappings       = 1.16.2-pre3+build.2
+minecraft_version   = 1.16.3
+yarn_mappings       = 1.16.3+build.46
 loader_version      = 0.1.0.48
 
 #Mod properties
