+++ src/main/java/com/remodstudios/endless_fabric/Endless.java
@@ -1,5 +1,6 @@
-package com.remodstudios.endless;
+package com.remodstudios.endless_fabric;
 
+import com.remodstudios.endless_fabric.common.register.EndlessModItems;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.minecraft.item.Item;
@@ -10,12 +11,11 @@ import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class endless implements ModInitializer {
-
-    public static Logger LOGGER = LogManager.getLogger();
-
+public class Endless implements ModInitializer {
     public static final String MOD_ID = "endless_fabric";
     public static final String MOD_NAME = "Endless";
+
+    public static Logger LOGGER = LogManager.getLogger(MOD_NAME);
 
     //item instances
     public static final Item TOPAZ_GEM = new Item(new FabricItemSettings().group(ItemGroup.MISC));
@@ -24,11 +24,12 @@ public class endless implements ModInitializer {
     public void onInitialize() {
         log(Level.INFO, "Initializing");
         // haha yes pichu smol and not best
-        Registry.register(Registry.ITEM, new Identifier("endless_fabric", "topaz_gem"), TOPAZ_GEM);
+        log(Level.INFO, "Registering Endless' Items...");
+        EndlessModItems.registerItems();
     }
 
     public static void log(Level level, String message){
-        LOGGER.log(level, "["+MOD_NAME+"] " + message);
+        LOGGER.log(level, message);
     }
 
 }
